# static/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "AI Interview Platform API",
    "description": "API documentation for the AI-Powered Technical Interview Platform",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/api",
      "description": "API Server"
    }
  ],
...
  "paths": {
...
    "/candidates": {
      "get": {
        "tags": ["Candidates"],
        "summary": "Get all candidates",
        "description": "Get a list of all candidate profiles",
        "parameters": [
          {
            "name": "skill",
            "in": "query",
            "description": "Filter by skill name",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "experience",
            "in": "query",
            "description": "Filter by minimum years of experience",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of results",
            "schema": {
              "type": "integer",
              "default": 20
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Pagination offset",
            "schema": {
              "type": "integer",
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of candidates",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "candidates": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "integer"
                          },
                          "first_name": {
                            "type": "string"
                          },
                          "last_name": {
                            "type": "string"
                          },
                          "preferred_role": {
                            "type": "string"
                          },
                          "experience_years": {
                            "type": "integer"
                          },
                          "skills": {
                            "type": "array",
                            "items": {
                              "type": "object",
                              "properties": {
                                "id": {
                                  "type": "integer"
                                },
                                "name": {
                                  "type": "string"
                                }
                              }
                            }
                          }
                        }
                      }
                    },
                    "total": {
                      "type": "integer"
                    },
                    "limit": {
                      "type": "integer"
                    },
                    "offset": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/candidates/{candidate_id}": {
      "get": {
        "tags": ["Candidates"],
        "summary": "Get candidate profile",
        "description": "Get detailed candidate profile by ID",
        "parameters": [
          {
            "name": "candidate_id",
            "in": "path",
            "required": true,
            "description": "ID of the candidate",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Candidate profile",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer"
                    },
                    "user_id": {
                      "type": "integer"
                    },
                    "first_name": {
                      "type": "string"
                    },
                    "last_name": {
                      "type": "string"
                    },
                    "full_name": {
                      "type": "string"
                    },
                    "phone": {
                      "type": "string"
                    },
                    "linkedin_url": {
                      "type": "string"
                    },
                    "github_url": {
                      "type": "string"
                    },
                    "resume_url": {
                      "type": "string"
                    },
                    "experience_years": {
                      "type": "integer"
                    },
                    "preferred_role": {
                      "type": "string"
                    },
                    "skills": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    },
                    "skill_assessments": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Candidate not found"
          }
        }
      }
    },
    "/candidates/profile": {
      "get": {
        "tags": ["Candidates"],
        "summary": "Get current candidate's profile",
        "description": "Get profile for the authenticated candidate",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Candidate profile",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "403": {
            "description": "Unauthorized access"
          },
          "404": {
            "description": "Profile not found"
          }
        }
      },
      "put": {
        "tags": ["Candidates"],
        "summary": "Update candidate profile",
        "description": "Update profile for the authenticated candidate",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "first_name": {
                    "type": "string"
                  },
                  "last_name": {
                    "type": "string"
                  },
                  "phone": {
                    "type": "string"
                  },
                  "linkedin_url": {
                    "type": "string"
                  },
                  "github_url": {
                    "type": "string"
                  },
                  "experience_years": {
                    "type": "integer"
                  },
                  "preferred_role": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Profile updated successfully"
          },
          "403": {
            "description": "Unauthorized access"
          },
          "404": {
            "description": "Profile not found"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}